refactor(app): extract useFileOps hook
